Refactored and got the Hotspot mode initialisation working
